getting ready to move types into specific folders
