+++ package.json
@@ -21,6 +21,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
+    "react-query": "^3.39.2",
     "react-router-dom": "^6.3.0",
     "sass": "^1.56.0",
     "typescript": "^4.7.4",
