--- src/components/Lists/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "../../pages/Profile/components/ProfileLists";
