+++ src/features/api/backendApi.ts
@@ -1,9 +1,10 @@
 import axios from "axios";
 
-const axiosInstance = axios.create({
-  baseURL: "https://dragpersonalproj.xyz/cinema-log",
+const backendApi = axios.create({
+  baseURL: "http://localhost:5001",
 });
 
+//handles preflight error
 axios.interceptors.response.use(null!, (error) => {
   if (error && error.message === "Network Error") {
     throw new Error(`Potential network CORS preflight error at ${error.config.url}`);
@@ -11,10 +12,10 @@ axios.interceptors.response.use(null!, (error) => {
   throw error;
 });
 
-axiosInstance.interceptors.request.use((config) => {
+backendApi.interceptors.request.use((config) => {
   const user = JSON.parse(localStorage.getItem("user")!);
   config.headers!["Authorization"] = `${user.username} ${user.token}`;
   return config;
 });
 
-export default axiosInstance;
+export default backendApi;
