+++ src/pages/Person/index.tsx
@@ -1,5 +1,5 @@
 import PersonActedIn from "./components/PersonActedIn";
-import { useFetch } from "../../hooks/useFetch";
+import { useFetch } from "@/hooks/useFetch";
 import { Container } from "@mui/system";
 import { useParams } from "react-router-dom";
 import { CardMedia, CircularProgress, Grid, Typography } from "@mui/material";
