--- src/pages/Register/styles/register.css
@@ -1,26 +1,0 @@
-.register-page {
-  display: flex;
-  justify-content: center;
-  background-size: cover;
-  height: 100vh;
-  color: #cccccc;
-}
-.register-page .form-container {
-  width: 65%;
-  background-color: black;
-  margin-right: 4.5rem;
-  height: 32rem;
-  margin-top: 0rem;
-}
-.register-page .form-container .registration-cont {
-  display: flex;
-  gap: 2rem;
-  justify-content: center;
-  align-items: center;
-}
-.register-page .form-container .registration-cont .landing-form > * {
-  font-family: nunito, sans-serif;
-  width: 20rem;
-  -webkit-animation: fade 500ms 1 ease-in-out forwards;
-          animation: fade 500ms 1 ease-in-out forwards;
-}/*# sourceMappingURL=register.css.map */
