--- src/styles/App.css
@@ -1,141 +1,0 @@
-@import url(https://fonts.googleapis.com/css2?family=Lato:wght@400;700&display=swap);
-@import url("https://use.typekit.net/yno2wnn.css");
-* {
-  box-sizing: border-box;
-  font-family: nunito, sans-serif;
-}
-
-a {
-  color: #fff;
-  text-decoration: none;
-}
-
-@-webkit-keyframes fade {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-@keyframes fade {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-html {
-  margin: 0;
-  padding: 0;
-}
-
-a {
-  text-decoration: none;
-}
-
-body {
-  margin: 0;
-  padding: 0;
-  background: #14181c;
-  min-height: 100vh;
-}
-body .main-page-cont {
-  display: grid;
-  height: 90vh;
-  justify-content: center;
-  grid-template-columns: 60% 40%;
-  grid-template-rows: repeat(auto-fill, 1fr);
-  color: #cccccc;
-  -webkit-animation: fade 500ms 1 ease-in-out forwards;
-          animation: fade 500ms 1 ease-in-out forwards;
-}
-body .main-page-cont .main-page-feed-cont {
-  grid-row: 1;
-  grid-column: 2;
-}
-body .main-page-cont .main-page-feed-cont .main-page-user-feed {
-  grid-column: 1/2;
-  padding: 1rem;
-  border-radius: 0.5rem;
-  background-color: #161b22;
-  width: 35rem;
-  white-space: nowrap;
-}
-body .main-page-cont .main-page-feed-cont .main-page-user-feed .main-page-user-feed-cont {
-  display: grid;
-  grid-template-columns: 70% 30%;
-  margin-bottom: 0.5em;
-  color: #cccccc;
-}
-body .main-page-cont .main-page-feed-cont .main-page-user-feed .main-page-user-feed-cont .user {
-  color: #fff;
-}
-body .main-page-cont .main-page-feed-cont .main-page-user-feed .main-page-user-feed-cont .user:hover {
-  cursor: pointer;
-}
-body .main-page-cont .main-page-list-container {
-  display: flex;
-  flex-direction: column;
-  gap: 0.3rem;
-  color: #cccccc;
-  background-color: #161b22;
-  padding-top: 1rem;
-  padding-left: 1rem;
-  padding-right: 1rem;
-  border-radius: 1em;
-}
-body .main-page-cont .main-page-list-container .list-created-by .list-username {
-  color: #fff;
-}
-body .main-page-cont .main-page-list-container .main-page-list-img-cont {
-  display: flex;
-}
-body .main-page-cont .main-page-list-container .main-page-list-img-cont .main-page-list-img {
-  width: 8em;
-}
-
-.img-form-menu:hover {
-  cursor: pointer;
-  background-color: #3a3a3a;
-}
-
-.movie-card-link-img {
-  border: 1px solid transparent;
-  transition: 0.25s border-color;
-  -webkit-user-select: none;
-     -moz-user-select: none;
-      -ms-user-select: none;
-          user-select: none;
-}
-.movie-card-link-img:hover {
-  border-color: #6d6d6d;
-}
-
-.svg:hover {
-  cursor: pointer;
-}
-
-@media (max-width: 1920px) {
-  body .cont .movie-main {
-    height: 24em;
-  }
-}
-@media (max-width: 2550px) {
-  .form-container {
-    height: 40rem;
-  }
-}
-@media (max-width: 1130px) {
-  .cont .movie-main img {
-    width: auto;
-  }
-}
-.spinner {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}/*# sourceMappingURL=App.css.map */
