--- src/styles/App.sass
@@ -1,126 +1,0 @@
-@import url(https://fonts.googleapis.com/css2?family=Lato:wght@400;700&display=swap)
-@import url("https://use.typekit.net/yno2wnn.css")
-
-$main: "Lato", sans-serif
-$main-color: #14181c
-$secondary-color: #161b22
-$Nunito: nunito, sans-serif
-$fade500ms: fade 500ms 1 ease-in-out forwards
-$notification: notification 1000ms 1 ease-in-out forwards
-
-@mixin flex()
-    display: flex
-    justify-content: center
-
-@mixin flexColumn()
-    display: flex
-    flex-direction: column
-
-*
-    box-sizing: border-box
-    font-family: $Nunito
-
-a
-    color: #fff
-    text-decoration: none
-
-@keyframes fade
-    0%
-        opacity: 0
-    100%
-        opacity: 1
-
-html
-    margin: 0
-    padding: 0
-
-a
-    text-decoration: none
-
-body
-    margin: 0
-    padding: 0
-    background: $main-color
-    min-height: 100vh
-    .main-page-cont
-        display: grid
-        height: 90vh
-        justify-content: center
-        grid-template-columns: 60% 40%
-        grid-template-rows: repeat(auto-fill, 1fr)
-        color: #cccccc
-        animation: $fade500ms
-        .main-page-feed-cont
-            grid-row: 1
-            grid-column: 2
-            .main-page-user-feed
-                grid-column: 1 / 2
-                padding: 1rem
-                border-radius: 0.5rem
-                background-color: $secondary-color
-                width: 35rem
-                white-space: nowrap
-                .main-page-user-feed-cont
-                    display: grid
-                    grid-template-columns: 70% 30%
-                    margin-bottom: 0.5em
-                    color: #cccccc
-                    .user
-                        color: #fff
-                        &:hover
-                            cursor: pointer
-        .main-page-list-container
-            @include flexColumn()
-            gap: 0.3rem
-            color: #cccccc
-            background-color: $secondary-color
-            padding-top: 1rem
-            padding-left: 1rem
-            padding-right: 1rem
-            border-radius: 1em
-            .list-created-by
-                .list-username
-                    color: #fff
-            .main-page-list-img-cont
-                display: flex
-                .main-page-list-img
-                    width: 8em
-
-.img-form-menu
-    &:hover
-        cursor: pointer
-        background-color: #3a3a3a
-
-.movie-card-link-img
-    border: 1px solid transparent
-    transition: 0.25s border-color
-    user-select: none
-    &:hover
-        border-color: #6d6d6d
-
-.svg
-    &:hover
-        cursor: pointer
-
-@media (max-width: 1920px)
-    body
-        .cont
-            .movie-main
-                height: 24em
-
-@media (max-width: 2550px)
-    .form-container
-        height: 40rem
-
-@media(max-width: 1130px)
-    .cont
-        .movie-main
-            img
-                width: auto
-
-//Centering spinner
-.spinner
-    position: absolute
-    top: 50%
-    left: 50%
-    transform: translate(-50%, -50%)
