+++ src/styles/_reset.sass
@@ -1,0 +1,60 @@
+@import "variables"
+
+.svg
+    &:hover
+        cursor: pointer
+
+*
+    box-sizing: border-box
+    font-family: $Nunito
+
+a
+    color: #fff
+    text-decoration: none
+
+body
+    margin: 0
+    padding: 0
+    background: $main-color
+    min-height: 100vh
+
+@keyframes fade
+    0%
+        opacity: 0
+    100%
+        opacity: 1
+
+.img-form-menu
+    &:hover
+        cursor: pointer
+        background-color: #3a3a3a
+
+.movie-card-link-img
+    border: 1px solid transparent
+    transition: 0.25s border-color
+    user-select: none
+    &:hover
+        border-color: #6d6d6d
+
+@media (max-width: 1920px)
+    body
+        .cont
+            .movie-main
+                height: 24em
+
+@media (max-width: 2550px)
+    .form-container
+        height: 40rem
+
+@media(max-width: 1130px)
+    .cont
+        .movie-main
+            img
+                width: auto
+
+//Centering spinner
+.spinner
+    position: absolute
+    top: 50%
+    left: 50%
+    transform: translate(-50%, -50%)
