+++ src/styles/components/_crew_cast.sass
@@ -1,4 +1,3 @@
-//Using this for actors too
 .cast-crew-cont
     display: grid
     grid-template-columns: repeat(2, 50%)
