+++ src/styles/components/_mediapagegenres.sass
@@ -11,6 +11,7 @@
         margin-bottom: 0.2em
         padding-bottom: 0.5em
         background-color: $main-color
+        border-radius: 0.5rem
         justify-content: center
         &:hover
             background-color: #171b20
