+++ src/styles/pages/_homepage.sass
@@ -1,0 +1,43 @@
+body
+    .main-page-cont
+        display: grid
+        height: 90vh
+        justify-content: center
+        grid-template-columns: 60% 40%
+        grid-template-rows: repeat(auto-fill, 1fr)
+        color: #cccccc
+        animation: $fade500ms
+        .main-page-feed-cont
+            grid-row: 1
+            grid-column: 2
+            .main-page-user-feed
+                grid-column: 1 / 2
+                padding: 1rem
+                border-radius: 0.5rem
+                background-color: $secondary-color
+                width: 35rem
+                white-space: nowrap
+                .main-page-user-feed-cont
+                    display: grid
+                    grid-template-columns: 70% 30%
+                    margin-bottom: 0.5em
+                    color: #cccccc
+                    .user
+                        color: #fff
+                        &:hover
+                            cursor: pointer
+            .main-page-list-container
+                gap: 0.3rem
+                color: #cccccc
+                background-color: $secondary-color
+                padding-top: 1rem
+                padding-left: 1rem
+                padding-right: 1rem
+                border-radius: 1em
+                .list-created-by
+                    .list-username
+                        color: #fff
+                .main-page-list-img-cont
+                    display: flex
+                    .main-page-list-img
+                        width: 8em
