+++ src/styles/pages/_person.sass
@@ -1,3 +1,5 @@
+@import "../variables"
+
 .person-cont
     margin-top: 5rem
     color: #cccccc
@@ -6,7 +8,7 @@
     .person-cards
         display: grid
         grid-template-columns: repeat(5, 13rem)
-        background-color: #161b22
+        background-color: $secondary-color
         justify-content: center
         padding-top: 1.5rem
         padding-bottom: 1.5rem
